requirements.txt
requests==2.25.0
python-telegram-bot==13.1
python-dotenv==0.15.0
google-cloud-dialogflow=2.0.0
vk-api=11.9.1